+++ src/ts/components/audioqualityselectbox.ts
@@ -15,15 +15,7 @@ export class AudioQualitySelectBox extends SelectBox {
     super.configure(player, uimanager);
 
     let selectCurrentAudioQuality = () => {
-      if (player.getAudioQuality) {
-        // Since player 7.3.1
-        this.selectItem(player.getAudioQuality().id);
-      } else {
-        // Backwards compatibility for players <= 7.3.0
-        // TODO remove in next major release
-        let data = player.getDownloadedAudioData();
-        this.selectItem(data.isAuto ? 'auto' : data.id);
-      }
+      this.selectItem(player.getAudioQuality().id);
     };
 
     let updateAudioQualities = () => {
@@ -48,21 +40,14 @@ export class AudioQualitySelectBox extends SelectBox {
     });
 
     // Update qualities when audio track has changed
-    player.addEventHandler(player.EVENT.ON_AUDIO_CHANGED, updateAudioQualities);
+    player.on(player.exports.Event.AudioChanged, updateAudioQualities);
     // Update qualities when source goes away
-    player.addEventHandler(player.EVENT.ON_SOURCE_UNLOADED, updateAudioQualities);
+    player.on(player.exports.Event.SourceUnloaded, updateAudioQualities);
     // Update qualities when a new source is loaded
-    player.addEventHandler(player.EVENT.ON_READY, updateAudioQualities);
+    player.on(player.exports.Event.Ready, updateAudioQualities);
     // Update qualities when the period within a source changes
-    player.addEventHandler(player.EVENT.ON_PERIOD_SWITCHED, updateAudioQualities);
+    player.on(player.exports.Event.PeriodSwitched, updateAudioQualities);
     // Update quality selection when quality is changed (from outside)
-    if (player.EVENT.ON_AUDIO_QUALITY_CHANGED) {
-      // Since player 7.3.1
-      player.addEventHandler(player.EVENT.ON_AUDIO_QUALITY_CHANGED, selectCurrentAudioQuality);
-    } else {
-      // Backwards compatibility for players <= 7.3.0
-      // TODO remove in next major release
-      player.addEventHandler(player.EVENT.ON_AUDIO_DOWNLOAD_QUALITY_CHANGE, selectCurrentAudioQuality);
-    }
+    player.on(player.exports.Event.AudioQualityChanged, selectCurrentAudioQuality);
   }
 }