+++ src/ts/components/metadatalabel.ts
@@ -62,8 +62,8 @@ export class MetadataLabel extends Label<MetadataLabelConfig> {
     // Init label
     init();
     // Reinit label when a new source is loaded
-    player.addEventHandler(player.EVENT.ON_SOURCE_LOADED, init);
+    player.on(player.exports.Event.SourceLoaded, init);
     // Clear labels when source is unloaded
-    player.addEventHandler(player.EVENT.ON_SOURCE_UNLOADED, unload);
+    player.on(player.exports.Event.SourceUnloaded, unload);
   }
 }