+++ src/ts/components/playbacktimelabel.ts
@@ -109,14 +109,13 @@ export class PlaybackTimeLabel extends Label<PlaybackTimeLabelConfig> {
       }
     };
 
-    player.addEventHandler(player.EVENT.ON_TIME_CHANGED, playbackTimeHandler);
-    player.addEventHandler(player.EVENT.ON_SEEKED, playbackTimeHandler);
-    player.addEventHandler(player.EVENT.ON_CAST_TIME_UPDATED, playbackTimeHandler);
+    player.on(player.exports.Event.TimeChanged, playbackTimeHandler);
+    player.on(player.exports.Event.Seeked, playbackTimeHandler);
 
-    player.addEventHandler(player.EVENT.ON_TIME_SHIFT, updateLiveTimeshiftState);
-    player.addEventHandler(player.EVENT.ON_TIME_SHIFTED, updateLiveTimeshiftState);
-    player.addEventHandler(player.EVENT.ON_PLAY, updateLiveTimeshiftState);
-    player.addEventHandler(player.EVENT.ON_PAUSED, updateLiveTimeshiftState);
+    player.on(player.exports.Event.TimeShift, updateLiveTimeshiftState);
+    player.on(player.exports.Event.TimeShifted, updateLiveTimeshiftState);
+    player.on(player.exports.Event.Play, updateLiveTimeshiftState);
+    player.on(player.exports.Event.Paused, updateLiveTimeshiftState);
 
     let init = () => {
       // Reset min-width when a new source is ready (especially for switching VOD/Live modes where the label content
@@ -133,7 +132,7 @@ export class PlaybackTimeLabel extends Label<PlaybackTimeLabelConfig> {
       // Update time after the format has been set
       playbackTimeHandler();
     };
-    player.addEventHandler(player.EVENT.ON_READY, init);
+    player.on(player.exports.Event.Ready, init);
 
     init();
   }