+++ src/ts/playerutils.ts
@@ -6,11 +6,11 @@ export namespace PlayerUtils {
   import PlayerAPI = bitmovin.PlayerAPI;
 
   export enum PlayerState {
-    IDLE,
-    PREPARED,
-    PLAYING,
-    PAUSED,
-    FINISHED,
+    Idle,
+    Prepared,
+    Playing,
+    Paused,
+    Finished,
   }
 
   export function isTimeShiftAvailable(player: bitmovin.PlayerAPI): boolean {
@@ -19,15 +19,15 @@ export namespace PlayerUtils {
 
   export function getState(player: PlayerAPI): PlayerState {
     if (player.hasEnded()) {
-      return PlayerState.FINISHED;
+      return PlayerState.Finished;
     } else if (player.isPlaying()) {
-      return PlayerState.PLAYING;
+      return PlayerState.Playing;
     } else if (player.isPaused()) {
-      return PlayerState.PAUSED;
-    } else if (player.isReady()) {
-      return PlayerState.PREPARED;
+      return PlayerState.Paused;
+    } else if (player.getConfig(true).source) {
+      return PlayerState.Prepared;
     } else {
-      return PlayerState.IDLE;
+      return PlayerState.Idle;
     }
   }
 
@@ -48,11 +48,11 @@ export namespace PlayerUtils {
       let timeShiftDetector = () => {
         this.detect();
       };
-      // Try to detect timeshift availability in ON_READY, which works for DASH streams
-      player.addEventHandler(player.EVENT.ON_READY, timeShiftDetector);
+      // Try to detect timeshift availability in Ready, which works for DASH streams
+      player.on(player.exports.Event.Ready, timeShiftDetector);
       // With HLS/NativePlayer streams, getMaxTimeShift can be 0 before the buffer fills, so we need to additionally
-      // check timeshift availability in ON_TIME_CHANGED
-      player.addEventHandler(player.EVENT.ON_TIME_CHANGED, timeShiftDetector);
+      // check timeshift availability in TimeChanged
+      player.on(player.exports.Event.TimeChanged, timeShiftDetector);
     }
 
     detect(): void {
@@ -102,15 +102,15 @@ export namespace PlayerUtils {
         this.detect();
       };
       // Initialize when player is ready
-      player.addEventHandler(player.EVENT.ON_READY, liveDetector);
+      player.on(player.exports.Event.Ready, liveDetector);
       // Re-evaluate when playback starts
-      player.addEventHandler(player.EVENT.ON_PLAY, liveDetector);
+      player.on(player.exports.Event.Play, liveDetector);
 
       // HLS live detection workaround for Android:
       // Also re-evaluate during playback, because that is when the live flag might change.
       // (Doing it only in Android Chrome saves unnecessary overhead on other plattforms)
       if (BrowserUtils.isAndroid && BrowserUtils.isChrome) {
-        player.addEventHandler(player.EVENT.ON_TIME_CHANGED, liveDetector);
+        player.on(player.exports.Event.TimeChanged, liveDetector);
       }
     }